requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
facenet-pytorch==2.6.0
fastapi==0.116.2
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.24.4
opencv-python==4.9.0.80
packaging==25.0
pillow==10.2.0
pydantic==2.11.9
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
